@@ -27,15 +27,15 @@
     "demo:release": "npm run demo:build && node scripts/gh-pages"
   },
   "dependencies": {
-    "@angular/animations": "^13.3.12",
+    "@angular/animations": "^14.2.12",
     "@angular/cdk": "^12.2.13",
-    "@angular/common": "^13.3.12",
-    "@angular/compiler": "^13.3.12",
-    "@angular/core": "^13.3.12",
-    "@angular/forms": "^13.3.12",
-    "@angular/platform-browser": "^13.3.12",
-    "@angular/platform-browser-dynamic": "^13.3.12",
-    "@angular/router": "^13.3.12",
+    "@angular/common": "^14.2.12",
+    "@angular/compiler": "^14.2.12",
+    "@angular/core": "^14.2.12",
+    "@angular/forms": "^14.2.12",
+    "@angular/platform-browser": "^14.2.12",
+    "@angular/platform-browser-dynamic": "^14.2.12",
+    "@angular/router": "^14.2.12",
     "cleave.js": "^1.6.0",
     "core-js": "^2.6.3",
     "date-fns": "^2.6.0",
@@ -46,10 +46,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.10",
-    "@angular/cli": "~13.3.10",
-    "@angular/compiler-cli": "^13.3.12",
-    "@angular/language-service": "^13.3.12",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "^14.2.12",
+    "@angular/language-service": "^14.2.12",
     "@salesforce-ux/design-system": "~2.9.0",
     "@stackblitz/sdk": "^1.3.0",
     "@types/jasmine": "~3.6.0",
@@ -76,7 +76,7 @@
     "karma-jasmine-html-reporter": "^1.5.0",
     "karma-sauce-launcher": "^2.0.2",
     "markdown-it": "^8.4.2",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "npm-run-all": "^4.1.5",
     "prismjs": "^1.23.0",
     "protractor": "~7.0.0",
