@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { Observable, of } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
 
@@ -18,7 +18,7 @@ const PARAMS = new HttpParams({
 })
 export class DemoLookupsHttp {
 
-  inputCtrl = new FormControl();
+  inputCtrl = new UntypedFormControl();
 
   wikiResults$: Observable<any[]>;
   wikiPage: string;
